product

+++ src/main/java/com/miu/waafinalproject/SecureOnlineAuctionSystem/repository/BidRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BidRepo extends JpaRepository<Bid, Long> {
+public interface BidRepo extends JpaRepository<Bid, Integer> {
 }
